Request enpoints creating/editing/getting (#20)
* request endpoints, luf git <3

* add swagger

+++ server/src/main/java/ovh/nixenos/tab/server/services/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
     @Autowired
     PasswordEncoder passwordEncoder;
 
-    public User findById(final int id) {
+    public User findById(final Long id) {
         User result = userRepository.findById(id);
         return result;
     }
